+++ react/__tests__/components/BuyButton.test.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { render, fireEvent } from '@vtex/test-tools/react'
-import { MockedProvider } from 'react-apollo/test-utils'
+import { MockedProvider } from '@apollo/react-testing'
 
 import BuyButton from '../../BuyButton'
 
@@ -11,13 +11,7 @@ describe('<BuyButton />', () => {
       ...customProps,
     }
 
-    const comp = (
-      <MockedProvider resolvers={{}}>
-        <BuyButton {...props}>{text}</BuyButton>
-      </MockedProvider>
-    )
-
-    return render(comp)
+    return render(<BuyButton {...props}>{text}</BuyButton>, { MockedProvider })
   }
 
   it('should be rendered', async () => {