+++ react/__tests__/components/SearchBar.test.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import { render } from '@vtex/test-tools/react'
+import { MockedProvider } from '@apollo/react-testing'
 
 import SearchBar from '../../SearchBar'
 import autocomplete from '../../components/SearchBar/queries/autocomplete.gql'
@@ -20,7 +21,7 @@ describe('<SearchBar />', () => {
   }
 
   const renderComponent = (customProps = {}) => {
-    return render(<SearchBar />, { graphql: { mocks: [mockedResult] } })
+    return render(<SearchBar />, { graphql: { mocks: [mockedResult] }, MockedProvider })
   }
 
   it('should be able to mount and not break', () => {